models/app/requirements.txt
streamlit>=1.30.0
wandb>=0.16.0
ultralytics>=8.0.0
Pillow>=10.0.0
